feat: make appointments, add product to cart, change quantity, remove product.

+++ backend/src/main/java/j2ee/j2ee/apps/service/ServiceService.java
@@ -1,7 +1,5 @@
 package j2ee.j2ee.apps.service;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -14,7 +12,7 @@ public class ServiceService {
     @Autowired
     private ServiceRepository serviceRepository;
 
-    public Page<ServiceEntity> getByStoreId(int storeId, int page, int size) {
+    public Page<ServiceEntity> getByStoreId(long storeId, int page, int size) {
         Pageable pageable = PageRequest.of(page, size, Sort.by("name"));
         return serviceRepository.findAllByStoreId(storeId, pageable);
     }
